show loading at least 500ms
